services_salvar-pedido/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.16.0
uuid
firebase-admin
flask